Add new unittests
